move apollo-extra logging classes into new package
