+++ apollo-http-service/src/main/java/com/spotify/apollo/httpservice/HttpServiceModule.java
@@ -26,7 +26,7 @@ import com.spotify.apollo.core.Service;
 import com.spotify.apollo.environment.ApolloEnvironment;
 import com.spotify.apollo.environment.ApolloEnvironmentModule;
 import com.spotify.apollo.environment.RequestRunnableFactoryDecorator;
-import com.spotify.apollo.logging.RequestLoggingDecorator;
+import com.spotify.apollo.logging.extra.RequestLoggingDecorator;
 import com.spotify.apollo.module.AbstractApolloModule;
 import com.spotify.apollo.request.RequestHandler;
